# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zoryqon API
  version: 1.0.0
  description: Public market data, authenticated customer trading, provider callbacks, and protected operations.
servers:
  - url: https://api.zoryqon.com
    description: Production
security: []
...
paths:
...
  /v1/auth/login:
    get:
      tags: [Identity]
      operationId: beginLogin
      parameters:
        - { name: returnTo, in: query, schema: { type: string, default: / } }
      responses:
        '302': { description: Redirect to the configured OIDC provider }
  /v1/auth/callback:
    get:
      tags: [Identity]
      operationId: completeLogin
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '302':
          { description: Secure cookies set and browser redirected to the validated return path }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token verified by issuer, audience, signature, and expiry.
    cookieAuth:
      type: apiKey
      in: cookie
      name: zoryqon_access
...